Refactored out IOMobileFramebuffer functions
